fix writing across drives, force  user to add proper path, initial playlist code

+++ archie/tasks/downloader.py
@@ -26,7 +26,8 @@ def init():
     db.Video.reset_download_states()
 
     # remove temp downloads (for safety, resuming causes issues sometimes)
-    shutil.rmtree(TEMP_DL_PATH)
+    if TEMP_DL_PATH.exists():
+        shutil.rmtree(TEMP_DL_PATH)
 
 
 download_counter = 0
@@ -61,7 +62,7 @@ def download_videos(config: Config):
                 archive_config = c
                 break
 
-        assert archive_config
+        assert archive_config, f"FATAL ERROR: archive {archive_name} no longer exists."
 
         log(f"downloading video {video.title} ({video.id})")
